Creation of default columns for Detailed View mode postponed a bit.
If widget is freshly created, it doesn't have any columns in view.
If we create some default columns then caller will want another
ones then the columns will be rearranged (which will do lot of work
inside of GtkTreeView). This is just a waste.
So creation of default columns is postponed to time before model is
bound to view - in that time if caller already set desired columns then
those columns will be used, otherwise default ones will be created.
No rearranges in any case.