Removed duplicate columns list from FmStandardView.
We already have a list of columns in GtkTreeView. If we have second
one in FmStandardView object we much probably will step onto problems
sooner or later due to some kind of unsync. That is very vulnerable
design therefore unacceptable for long running application.

The secondary list replaced. The objects from list are attached to
columns itself as fm_quark_id data and only GtkTreeView internal list
is used as columns list from now on to exclude duplications and syncs.