Use FmPath to check monitoring events since some VFS may have mixed dirs.
For example, search:// may return few results with the same basename so we
cannot handle that case. Therefore fm_path_compare and fm_path_equal should
be used to looking and FmPath should be kept in lists. That will also take
less memory due to no strdup() but fm_path_ref() is used.